Make warnings in rusty-cpp headers visible during testing

+++ test_package/src/test_package.cpp
@@ -7,10 +7,10 @@
 TEST_F(Test, Asserts) {
 	rusty_assert(1);
 	rusty_assert_eq(
-		rusty::next_power_of_two((size_t)42), 64, "Math doesn't exist anymore"
+		rusty::next_power_of_two((size_t)42), (size_t)64, "Math doesn't exist anymore"
 	);
-	rusty_assert_eq(rusty::next_power_of_two((size_t)233), 256);
-	rusty_assert_ne(rusty::next_power_of_two((size_t)233), 255);
+	rusty_assert_eq(rusty::next_power_of_two((size_t)233), (size_t)256);
+	rusty_assert_ne(rusty::next_power_of_two((size_t)233), (size_t)255);
 }
 
 int main(int argc, char **argv) {
